+++ changelog.md
@@ -1,3 +1,11 @@
 January 12 2026:
     -Created proposal
     -Some code has already been written at home
+
+January 13 2026 
+- Added phrases
+- Added some code for randomly picking a phrase
+
+January 14
+-Player can now pick letters
+-Fixed capitalization on phrases